requirements.txt
python-dotenv==1.0.1
beautifulsoup4==4.12.3
chardet==5.2.0
requests==2.32.3
charset-normalizer==3.4.0
colorama==0.4.6